Fixed some compilation warning on x86 32-bit hosts and GCC 8.4.0.
Former-commit-id: 6d9d4ce7ee159f33b7537de0e1bcc282a85a9cb3

+++ unisim_lib/unisim/service/web_terminal/web_terminal.cc
@@ -499,7 +499,7 @@ ScreenBufferLine& ScreenBuffer::GetLine(unsigned int line_index)
 bool ScreenBuffer::IsAtCursorLinePos(const_iterator it) const
 {
 	unsigned int cursor_line_index = height - display_height + cursor_lineno - 1;
-	return (it - Begin()) == cursor_line_index;
+	return (unsigned int)(it - Begin()) == cursor_line_index;
 }
 
 ScreenBufferLine& ScreenBuffer::operator [] (unsigned int lineno)
